Chat: moved the loading of patient data to the ChatDetails component.
